+++ EstimoteSDK/EstimoteSDK.framework/Versions/A/Headers/ESTDeviceSettingProtocol.h
@@ -12,6 +12,8 @@
 #import <Foundation/Foundation.h>
 #import "ESTSettingProtocol.h"
 
+NS_ASSUME_NONNULL_BEGIN
+
 
 @protocol ESTDeviceSettingProtocol <ESTSettingProtocol>
 
@@ -25,12 +27,13 @@
 - (uint16_t)registerID;
 
 /**
- *  Method returns value converted to NSDate object that can be used
- *  to save value to the device.
+ *  Method returns value converted to NSData object that can be used
+ *  to save value to the device. For read-only settings (temperature)
+ *  this method can return nil.
  *
- *  @return NSDate representing setting value
+ *  @return NSData representing setting value
  */
-- (NSData *)getValueData;
+- (NSData * _Nullable)getValueData;
 
 /**
  *  Method updates setting value based on NSData object delivered from the device.
@@ -61,3 +64,5 @@
 - (NSError *)validateValue;
 
 @end
+
+NS_ASSUME_NONNULL_END