+++ EstimoteSDK/EstimoteSDK.framework/Versions/A/Headers/ESTNearableSettingBroadcastingPower.h
@@ -14,6 +14,8 @@
 #import "ESTCloudSettingProtocol.h"
 #import "ESTBeaconDefinitions.h"
 
+NS_ASSUME_NONNULL_BEGIN
+
 
 @interface ESTNearableSettingBroadcastingPower : NSObject <ESTDeviceNearableSettingProtocol, ESTCloudSettingProtocol, NSCopying>
 
@@ -23,3 +25,5 @@
                    completion:(ESTSettingBroadcastingPowerCompletionBlock)completion;
 
 @end
+
+NS_ASSUME_NONNULL_END