+++ EstimoteSDK/EstimoteSDK.framework/Versions/A/Headers/EstimoteSDK.h
@@ -7,7 +7,7 @@
 //  |______|___/\__|_|_| |_| |_|\___/ \__\___| |_____/|_____/|_|\_\
 //
 //
-//  Version: 3.7.1
+//  Version: 3.8.0
 //  Copyright (c) 2015 Estimote. All rights reserved.
 
 /**
@@ -79,6 +79,7 @@
 #import "ESTRequestGetBeacons.h"
 #import "ESTRequestBeaconDetails.h"
 #import "ESTRequestGetBeaconsDetails.h"
+#import "ESTRequestGetBeaconsPublicDetails.h"
 #import "ESTRequestBeaconColor.h"
 #import "ESTRequestBeaconMac.h"
 #import "ESTRequestAssignGPSLocation.h"